1) Converted all long indices to int. It turns out, we cannot have arrays with more elements than int.MaxValue either way. 2) All strategy objects (recombination, mutation, selection) in GAs now use I
Troschuetz.Random.IGenerator objects for the stochastic operators. The user can pass in a IGenerator object in the constructors. Otherwise a default one will be used, which is a static field of Optimization/Commons/RandomNumberGenerationUtilities and is the same for the entire optimization process.

+++ ISAAR.MSolve.Analyzers/Optimization/Commons/BinaryUtilities.cs
@@ -15,10 +15,10 @@ namespace ISAAR.MSolve.Analyzers.Optimization.Commons
         /// <param name="start"></param>
         /// <param name="end">Exclusive</param>
         /// <returns></returns>
-        public static int BinaryToDecimal(bool[] bits, long start, long length)
+        public static int BinaryToDecimal(bool[] bits, int start, int length)
         {
             int dec = 0;
-            for (long i = start; i < start + length; ++i)
+            for (int i = start; i < start + length; ++i)
             {
                 dec += bits[i] ? dec + 1: dec;
                 //Console.WriteLine(dec);
@@ -33,11 +33,11 @@ namespace ISAAR.MSolve.Analyzers.Optimization.Commons
         /// <param name="start"></param>
         /// <param name="end">Exclusive</param>
         /// <returns></returns>
-        public static int GrayCodeToDecimal(bool[] bits, long start, long length)
+        public static int GrayCodeToDecimal(bool[] bits, int start, int length)
         {
             bool bin = bits[start];
             int dec = bin ? 1 : 0;
-            for (long i = start + 1; i < start + length; ++i) // n-1 repetitions
+            for (int i = start + 1; i < start + length; ++i) // n-1 repetitions
             {
                 bin = bin != bits[i];
                 dec += bin ? dec + 1 : dec;
